+++ api-types/src/vm_config/mod.rs
@@ -5,4 +5,5 @@
 
 pub(crate) mod cpus_config;
 pub(crate) mod memory_config;
+pub(crate) mod net_config;
 pub(crate) mod numa_config;
