+++ api-types/src/vm_config/net_config.rs
@@ -1,0 +1,32 @@
+// Copyright © 2026 Contributors to the Cloud Hypervisor project
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
+pub enum VhostMode {
+    #[default]
+    Client,
+    Server,
+}
+
+#[derive(Debug)]
+pub enum ParseVhostModeError {
+    InvalidValue(String),
+}
+
+impl FromStr for VhostMode {
+    type Err = ParseVhostModeError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "client" => Ok(VhostMode::Client),
+            "server" => Ok(VhostMode::Server),
+            _ => Err(ParseVhostModeError::InvalidValue(s.to_owned())),
+        }
+    }
+}
